DIAC-2305
