--- src/main/java/uk/gov/hmcts/reform/iacasedocumentsapi/domain/handlers/presubmit/letter/AbstractInternalCmrListingLetterTemplate.java
@@ -1,81 +1,0 @@
-package uk.gov.hmcts.reform.iacasedocumentsapi.domain.handlers.presubmit.letter;
-
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.AsylumCase;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.HearingCentre;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.ccd.CaseDetails;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.service.StringProvider;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.templates.DocumentTemplate;
-import uk.gov.hmcts.reform.iacasedocumentsapi.infrastructure.CustomerServicesProvider;
-
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.AsylumCaseDefinition.CMR_HEARING_CENTRE;
-import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.AsylumCaseDefinition.CMR_HEARING_DATE;
-import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.utils.AsylumCaseUtils.formatDateForRendering;
-import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.utils.AsylumCaseUtils.formatDateTimeForRendering;
-import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.utils.AsylumCaseUtils.getAppellantPersonalisation;
-import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.utils.AsylumCaseUtils.getCmrHearingChannel;
-
-/**
- * Common field mapping for the letters generated when a case management review hearing is listed.
- * Subclasses only supply the address the letter is sent to.
- */
-public abstract class AbstractInternalCmrListingLetterTemplate implements DocumentTemplate<AsylumCase> {
-
-    private static final DateTimeFormatter DOCUMENT_DATE_FORMAT = DateTimeFormatter.ofPattern("d MMMM yyyy");
-    private static final DateTimeFormatter DOCUMENT_TIME_FORMAT = DateTimeFormatter.ofPattern("HHmm");
-
-    private final String templateName;
-    private final CustomerServicesProvider customerServicesProvider;
-    private final StringProvider stringProvider;
-
-    protected AbstractInternalCmrListingLetterTemplate(
-        String templateName,
-        CustomerServicesProvider customerServicesProvider,
-        StringProvider stringProvider
-    ) {
-        this.templateName = templateName;
-        this.customerServicesProvider = customerServicesProvider;
-        this.stringProvider = stringProvider;
-    }
-
-    @Override
-    public String getName() {
-        return templateName;
-    }
-
-    @Override
-    public Map<String, Object> mapFieldValues(
-        CaseDetails<AsylumCase> caseDetails
-    ) {
-        final AsylumCase asylumCase = caseDetails.getCaseData();
-        final HearingCentre listedHearingCentre =
-            asylumCase
-                .read(CMR_HEARING_CENTRE, HearingCentre.class)
-                .orElseThrow(() -> new IllegalStateException("cmrHearingCentre is not present"));
-
-        final Map<String, Object> fieldValues = new HashMap<>();
-
-        fieldValues.putAll(getAppellantPersonalisation(asylumCase));
-        fieldValues.put("customerServicesTelephone", customerServicesProvider.getInternalCustomerServicesTelephone(asylumCase));
-        fieldValues.put("customerServicesEmail", customerServicesProvider.getInternalCustomerServicesEmail(asylumCase));
-        fieldValues.put("hearingLocation", stringProvider.get("hearingCentreAddress", listedHearingCentre.toString()).orElse("").replaceAll(",\\s*", "\n"));
-        fieldValues.put("hearingDate", formatDateTimeForRendering(asylumCase.read(CMR_HEARING_DATE, String.class).orElse(""), DOCUMENT_DATE_FORMAT));
-        fieldValues.put("hearingTime", formatDateTimeForRendering(asylumCase.read(CMR_HEARING_DATE, String.class).orElse(""), DOCUMENT_TIME_FORMAT));
-        fieldValues.put("dateLetterSent", formatDateForRendering(LocalDate.now().toString(), DOCUMENT_DATE_FORMAT));
-        fieldValues.put("hearingChannel", getCmrHearingChannel(asylumCase, "Unknown"));
-
-        List<String> recipientAddress = getRecipientAddress(asylumCase);
-
-        for (int i = 0; i < recipientAddress.size(); i++) {
-            fieldValues.put("address_line_" + (i + 1), recipientAddress.get(i));
-        }
-        return fieldValues;
-    }
-
-    protected abstract List<String> getRecipientAddress(AsylumCase asylumCase);
-}
