+++ src/test/java/uk/gov/hmcts/reform/iacasedocumentsapi/domain/entities/DocumentTagTest.java
@@ -99,8 +99,6 @@ class DocumentTagTest {
         assertEquals("internalCmrListingLetterBundle", DocumentTag.INTERNAL_CMR_LISTING_LETTER_BUNDLE.toString());
         assertEquals("internalCmrListingLrLetter", DocumentTag.INTERNAL_CMR_LISTING_LR_LETTER.toString());
         assertEquals("internalCmrListingLrLetterBundle", DocumentTag.INTERNAL_CMR_LISTING_LR_LETTER_BUNDLE.toString());
-        assertEquals("internalCmrListingAppellantLetter", DocumentTag.INTERNAL_CMR_LISTING_APPELLANT_LETTER.toString());
-        assertEquals("internalCmrListingAppellantLetterBundle", DocumentTag.INTERNAL_CMR_LISTING_APPELLANT_LETTER_BUNDLE.toString());
         assertEquals("iAUT2Form", DocumentTag.IAUT_2_FORM.toString());
         assertEquals("", DocumentTag.NONE.toString());
         assertEquals("internalOutOfTimeDecisionLetter", DocumentTag.INTERNAL_OUT_OF_TIME_DECISION_LETTER.toString());
@@ -135,6 +133,6 @@ class DocumentTagTest {
 
     @Test
     public void if_this_test_fails_it_is_because_it_needs_updating_with_your_changes() {
-        assertEquals(132, DocumentTag.values().length);
+        assertEquals(130, DocumentTag.values().length);
     }
 }
