+++ src/test/java/uk/gov/hmcts/reform/iacasedocumentsapi/infrastructure/config/DocumentCreatorConfigurationTest.java
@@ -13,9 +13,7 @@ import uk.gov.hmcts.reform.iacasedocumentsapi.domain.service.DocumentCreator;
 import uk.gov.hmcts.reform.iacasedocumentsapi.domain.service.DocumentGenerator;
 import uk.gov.hmcts.reform.iacasedocumentsapi.domain.service.DocumentUploader;
 import uk.gov.hmcts.reform.iacasedocumentsapi.domain.templates.CmrHearingNoticeTemplate;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.handlers.presubmit.letter.InternalCmrListingNonDetainedAppellantLetterTemplate;
 import uk.gov.hmcts.reform.iacasedocumentsapi.domain.handlers.presubmit.letter.InternalCmrListingLetterTemplate;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.handlers.presubmit.letter.InternalCmrListingNonDetainedOrDetainedInPrisonOrIrcLrLetterTemplate;
 import uk.gov.hmcts.reform.iacasedocumentsapi.domain.handlers.presubmit.letter.InternalCmrReListingLetterTemplate;
 import uk.gov.hmcts.reform.iacasedocumentsapi.domain.templates.letter.InternalDetainedCmrListingLetterTemplate;
 
@@ -26,8 +24,6 @@ class DocumentCreatorConfigurationTest {
     @Mock private CmrHearingNoticeTemplate cmrHearingNoticeTemplate;
     @Mock private InternalDetainedCmrListingLetterTemplate internalDetainedCmrListingLetterTemplate;
     @Mock private InternalCmrListingLetterTemplate internalCmrListingLetterTemplate;
-    @Mock private InternalCmrListingNonDetainedOrDetainedInPrisonOrIrcLrLetterTemplate internalCmrListingNonDetainedOrDetainedInPrisonOrIrcLrLetterTemplate;
-    @Mock private InternalCmrListingNonDetainedAppellantLetterTemplate internalCmrListingNonDetainedAppellantLetterTemplate;
     @Mock private InternalCmrReListingLetterTemplate internalCmrReListingLetterTemplate;
     @Mock private DocumentGenerator documentGenerator;
     @Mock private DocumentUploader documentUploader;
@@ -100,38 +96,6 @@ class DocumentCreatorConfigurationTest {
                 fileName,
                 fileNameQualifier,
                 internalCmrListingLetterTemplate,
-                documentGenerator,
-                documentUploader
-            );
-
-        assertThat(documentCreator).isNotNull();
-    }
-
-    @Test
-    void should_create_internal_cmr_listing_lr_letter_document_creator() {
-        DocumentCreator<AsylumCase> documentCreator =
-            configuration.getInternalCmrListingLrLetterDocumentCreator(
-                contentType,
-                fileExtension,
-                fileName,
-                fileNameQualifier,
-                    internalCmrListingNonDetainedOrDetainedInPrisonOrIrcLrLetterTemplate,
-                documentGenerator,
-                documentUploader
-            );
-
-        assertThat(documentCreator).isNotNull();
-    }
-
-    @Test
-    void should_create_internal_cmr_listing_appellant_letter_document_creator() {
-        DocumentCreator<AsylumCase> documentCreator =
-            configuration.getInternalCmrListingAppellantLetterDocumentCreator(
-                contentType,
-                fileExtension,
-                fileName,
-                fileNameQualifier,
-                    internalCmrListingNonDetainedAppellantLetterTemplate,
                 documentGenerator,
                 documentUploader
             );
